requirements.txt
numpy
faiss-cpu
rank-bm25
pypdf
sentence-transformers
torch
transformers
tqdm
sounddevice
pyttsx3
scikit-learn
sentence-transformers
transformers
torch
faiss-cpu
rank-bm25
tqdm
pypdf
